# Speedup wire_map in default.qubit device