--- a/pennylane/_qubit_device.py
+++ b/pennylane/_qubit_device.py
@@ -609,7 +609,7 @@ class QubitDevice(Device):
         Args:
             prob: The probabilities to return the marginal probabilities
                 for
-            wires (Iterable): wire labels for which to return
+            wires (Iterable[Number, str], Number, str, Wires): wires to return
                 marginal probabilities for. Wires not provided
                 are traced out of the system.