--- a/pennylane/devices/tests/test_wires.py
+++ b/pennylane/devices/tests/test_wires.py
@@ -63,4 +63,3 @@ class TestWiresIntegration:
         circuit2 = circuit_factory(dev2, wires2)
 
         assert np.allclose(circuit1(), circuit2(), atol=tol(dev1.analytic))
-