To reduce the time PennyLane spends in the `Wires` class further, we can optimise the process of converting the user's wire labels to the ones recognised by the device. The problem is that this would require a refactor of all devices. Instead, this PR overwrites some functions in `DefaultQubit` to only implement the change for the default.qubit device. The changes lead to a significant speedup.

**Details:**

The wire translation is handled in various places in the device by the line: 
``` python
device_wires = self.map_wires(op.wires)
```
`self.map_wires` calls the `Wires` class `map` method:
``` python
mapped_wires = wires.map(self.wire_map)
```
which does the translation via
```python
new_wires = [wire_map[w] for w in self]
```
Unfortunately, even if we change the `self.wire_map` function to not return `Wires` objects, such a change would be overwritten by `Wires.map` (which parses the output to `Wires` anyways).

In this PR, we overwrite `self.wire_map` and `self.map_wires` of the default.qubit device to completely bypass the `Wires.map` function (which is possibly a better design anyways). 

**Issues:**
Since default.qubit inherits and uses the `samples` and `marginal_probs` methods from `QubitDevice` (which we don't want to touch due to other devices inheriting from it), we need to insert one trivial change and one `if` statement into `QubitDevice`.

**Speedup:**

<pre>       before           after         ratio
     [ca35ffe6]       [eed5c1bf]
     &lt;master&gt;         &lt;speedup_wires_map_defaultqubit&gt;
       7.38±0.8ms       6.93±0.9ms     0.94  asv.core_suite.CircuitEvaluation.time_circuit(10, 3)
         13.6±2ms         12.8±2ms     0.94  asv.core_suite.CircuitEvaluation.time_circuit(10, 6)
         17.6±2ms         17.8±2ms     1.01  asv.core_suite.CircuitEvaluation.time_circuit(10, 9)
       2.37±0.2ms       2.20±0.2ms     0.93  asv.core_suite.CircuitEvaluation.time_circuit(2, 3)
       3.59±0.4ms       3.30±0.4ms     0.92  asv.core_suite.CircuitEvaluation.time_circuit(2, 6)
       5.59±0.9ms       4.27±0.4ms    ~0.76  asv.core_suite.CircuitEvaluation.time_circuit(2, 9)
       4.17±0.4ms       3.52±0.3ms    ~0.84  asv.core_suite.CircuitEvaluation.time_circuit(5, 3)
       7.12±0.8ms       6.27±0.8ms    ~0.88  asv.core_suite.CircuitEvaluation.time_circuit(5, 6)
         9.14±1ms         8.08±1ms    ~0.88  asv.core_suite.CircuitEvaluation.time_circuit(5, 9)
         10.3±1ms         9.61±1ms     0.93  asv.core_suite.GradientComputation.time_gradient_autograd(2, 3)
         28.2±2ms         22.8±3ms    ~0.81  asv.core_suite.GradientComputation.time_gradient_autograd(2, 6)
         45.8±2ms         40.6±2ms    ~0.89  asv.core_suite.GradientComputation.time_gradient_autograd(5, 3)
          152±2ms          143±2ms     0.94  asv.core_suite.GradientComputation.time_gradient_autograd(5, 6)
       12.6±0.8ms       11.2±0.3ms    ~0.89  asv.core_suite.GradientComputation.time_gradient_tf(2, 3)
<font color="#4E9A06">-        30.1±1ms       26.7±0.3ms     0.89  </font>asv.core_suite.GradientComputation.time_gradient_tf(2, 6)
       49.9±0.3ms         46.4±1ms     0.93  asv.core_suite.GradientComputation.time_gradient_tf(5, 3)
          167±7ms          154±2ms     0.92  asv.core_suite.GradientComputation.time_gradient_tf(5, 6)
       11.5±0.7ms         8.02±1ms    ~0.70  asv.core_suite.GradientComputation.time_gradient_torch(2, 3)
         24.8±2ms       21.4±0.6ms    ~0.86  asv.core_suite.GradientComputation.time_gradient_torch(2, 6)
<font color="#4E9A06">-        46.9±1ms       39.9±0.5ms     0.85  </font>asv.core_suite.GradientComputation.time_gradient_torch(5, 3)
          154±2ms          141±2ms     0.91  asv.core_suite.GradientComputation.time_gradient_torch(5, 6)
<font color="#4E9A06">-      1.01±0.04s          898±6ms     0.89  </font>asv.core_suite.Optimization.time_optimization_autograd
       1.11±0.03s          995±7ms    ~0.90  asv.core_suite.Optimization.time_optimization_tf
<font color="#4E9A06">-      1.02±0.04s          895±9ms     0.88  </font>asv.core_suite.Optimization.time_optimization_torch
</pre>